`ft_strlen` returns `size_t`

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: ekorley <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/23 16:39:03 by ekorley           #+#    #+#             */
-/*   Updated: 2016/09/23 18:31:54 by ekorley          ###   ########.fr       */
+/*   Updated: 2016/09/24 09:35:02 by ekorley          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *s)
+size_t		ft_strlen(const char *s)
 {
-	unsigned int	count;
+	size_t		count;
 
 	count = 0;
 	while (s != NULL)
